Bloom tweaks

+++ examples/stress_tests/many_foxes.rs
@@ -131,7 +131,7 @@ fn setup(
             commands.entity(ring_parent).with_children(|builder| {
                 builder.spawn(SceneBundle {
                     scene: fox_handle.clone(),
-                    transform: Transform::from_xyz(x as f32, 0.0, z as f32)
+                    transform: Transform::from_xyz(x, 0.0, z)
                         .with_scale(Vec3::splat(0.01))
                         .with_rotation(base_rotation * Quat::from_rotation_y(-fox_angle)),
                     ..default()
